Rename FiltersState to PageFilters to ease reviewing and fix tests
