+++ frontend/lib/providers/page_filters.dart
@@ -14,54 +14,106 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+import 'package:dartx/dartx.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../models/filters.dart';
+import '../filtering/artefact_filters.dart';
+import '../filtering/enriched_test_execution_filters.dart';
 import '../routing.dart';
-import 'artefact_environments.dart';
+import 'enriched_test_executions.dart';
 import 'family_artefacts.dart';
+import '../filtering/multi_option_filter.dart';
 
 part 'page_filters.g.dart';
+
+typedef FilterOptionState = ({String name, bool isSelected});
+
+typedef FilterState = ({String name, List<FilterOptionState> options});
 
 @riverpod
 class PageFilters extends _$PageFilters {
   @override
-  Filters build(Uri pageUri) {
+  List<FilterState> build(Uri pageUri) {
+    final queryParams = pageUri.queryParametersAll;
     if (AppRoutes.isDashboardPage(pageUri)) {
       final family = AppRoutes.familyFromUri(pageUri);
-      final artefacts = ref
-          .watch(familyArtefactsProvider(family))
-          .requireValue
-          .values
-          .toList();
-
-      return createEmptyArtefactFilters(family)
-          .copyWithOptionsExtracted(artefacts)
-          .copyWithQueryParams(pageUri.queryParametersAll);
+      final artefacts = ref.watch(familyArtefactsProvider(family)).value ?? {};
+      return _createFiltersState(
+        getArtefactFiltersFor(family),
+        queryParams,
+        artefacts.values.toList(),
+      );
     }
 
     if (AppRoutes.isArtefactPage(pageUri)) {
       final artefactId = AppRoutes.artefactIdFromUri(pageUri);
-      final artefactEnvironments =
-          ref.watch(artefactEnvironmentsProvider(artefactId)).requireValue;
-
-      return emptyArtefactEnvironmentsFilters
-          .copyWithOptionsExtracted(artefactEnvironments)
-          .copyWithQueryParams(pageUri.queryParametersAll);
+      final enrichedExecutions =
+          ref.watch(enrichedTestExecutionsProvider(artefactId)).value ?? [];
+      return _createFiltersState(
+        enrichedTestExecutionFilters,
+        queryParams,
+        enrichedExecutions,
+      );
     }
 
-    throw Exception('Called pageFiltersProvider in unknown page $pageUri');
+    throw Exception('Called filtersStateProvider in unknown page $pageUri');
   }
 
-  void handleFilterOptionChange(
+  List<FilterState> _createFiltersState<T>(
+    List<MultiOptionFilter<T>> filters,
+    Map<String, List<String>> queryParams,
+    List<T> items,
+  ) {
+    final result = <FilterState>[];
+    for (var filter in filters) {
+      final selectedOptions = (queryParams[filter.name] ?? []).toSet();
+      final allOptions = filter.extractOptions(items);
+      result.add(
+        (
+          name: filter.name,
+          options: allOptions
+              .map(
+                (option) => (
+                  name: option,
+                  isSelected: selectedOptions.contains(option),
+                ),
+              )
+              .toList()
+        ),
+      );
+    }
+    return result;
+  }
+
+  void onChanged(
     String filterName,
     String optionName,
-    bool optionValue,
+    bool isSelected,
   ) {
-    state = state.copyWithFilterOptionValue(
-      filterName,
-      optionName,
-      optionValue,
-    );
+    state = state.map((filter) {
+      if (filter.name != filterName) {
+        return filter;
+      }
+      return (
+        name: filterName,
+        options: filter.options
+            .map(
+              (option) => (
+                name: option.name,
+                isSelected:
+                    option.name == optionName ? isSelected : option.isSelected
+              ),
+            )
+            .toList(),
+      );
+    }).toList();
   }
+
+  Map<String, List<String>> toQueryParams() => ({
+        for (var filter in state)
+          filter.name: filter.options
+              .filter((option) => option.isSelected)
+              .map((option) => option.name)
+              .toList(),
+      });
 }
